Set up basic CI workflow
This is the same basic CI workflow we would expect to see for any Gradle
project: build the project and run the tests for every PR.  Exactly
which tests are run as part of the `check` task is left to
`build.gradle.kts`; doing so minimizes the amount of build code that
only runs on CI/CD and keeps the development feedback loop short.

Rename the "Hello, World" task from `run` to `check`, so that when we
evolve the project into a set of Gradle plugins we won't need to modify
the CI/CD workflow.

Defer the "CD" part of CI/CD until we have something that can be
meaningfully deployed.