Skeleton of Kotlin convention plugin
Create a new convention plugin for arbitrary Kotlin projects.  This
initial version simply wraps the standard Kotlin language plugin (and
therefore encapsulates the Kotlin language version to use); I will
first sort out concerns like publishing for cross-repository usage, and
then expand it to include things like static analysis tools.

Since I will be publishing this plugin to GitHub Packages, it makes
sense to use `io.github.aaron-harris` as the base for my Gradle
coordinates, both for the plugin itself and for the group identifier of
its implementation artifact.  Note that because my GitHub username
includes a hyphen, the "path" part of the plugin identifier cannot be
expressed as a JVM package, so in order to get the right ID I needed to
include the whole thing in the filename.

The actual configuration to publish the plugin to GitHub Packages is not
present yet, but by publishing it to my Maven Local repository, I was
able to confirm that it is built correctly and can be consumed by
external projects under the expected identifier.