Configure Detekt for plugin build
Add configuration to use the Detekt static analysis plugin for the build
of the `aph-kotlin` plugin itself.  As with Ktlint, this applies both to
the plugin source code (under `aph-kotlin/src`) and to the plugin's own
build script (`aph-kotlin/build.gradle.kts`).  It does not apply to
downstream projects that consume the plugin, nor to the root module's
build scripts (such as `settings.gradle.kts`).

The Detekt rules configuration was copied directly from my
[`pew-kotlin` project](https://github.com/aaron-harris/pew-kotlin/blob/36838db1a190f99a8f7815add102055daeee354e/config/detekt/detekt.yaml).