Configure Ktlint in `aph-kotlin` plugin
Configure Ktlint as a linter again, but this time for downstream
consumers of the `aph-kotlin` plugin.

The version of Ktlint is currently hard-coded at 1.3.0.  This is the
same version used in the plugin build itself, but because it needs to be
specified in the plugin source itself, linking the two versions is
challenging.  We can't reference the version catalog here because the
catalog resolved will be that of the consuming project, not this one.
The best long-term approach may be to leave the two references separate,
but add a test to verify that they remain in sync, as discussed in
[this blog post](https://lukeplant.me.uk/blog/posts/keeping-things-in-sync-derive-vs-test/).
That will entail learning Gradle's plugin-testing tools, though, so
let's leave that alone for now.

More specific Ktlint configuration (such as disabled rules) is governed
by the consuming project's `.editorconfig` file and is not specified
here.

Note that I am avoiding use of `spotless` as a type-safe accessor in the
plugin source here, using `configure<SpotlessExtension>` instead.  This
is because my IDE (IntelliJ 2024.1.4) can't recognize `spotless` without
an explicit import of the generated accessor, which should not be
necessary.  Gradle itself compiles the plugin fine, so I believe this is
an IDE bug only, but better to avoid type-safe accessors until it is
fixed.