Add configuration for unit tests and code coverage
Include my default configuration for unit tests (mostly concerned with
ensuring that unit test output is visible in the console), and add
configuration to use the Kover plugin for measuring code coverage.

Since we expect consuming projects to need to set their own coverage
thresholds, do not include that configuration yet.  The next step is to
create a plugin extension for `aph-kotlin` that will include that as a
configurable property.  This will allow us to set a default in a way
that can be overridden by consuming projects, and without the consuming
project having any knowledge of Kover's configuration DSL.