Add DSL extension for the `aph-kotlin` plugin
This extension currently contains a single configurable property, for
the minimum code coverage threshold that should be required.

The point of making this configurable is to allow consuming projects to
set the threshold without needing to know about the details of how to do
so using Kover's DSL, or in fact even whether code coverage is being
measured by Kover or Jacoco.  Also, I'm pretty sure that Kover's
verification rules are additive, meaning that the alternative approach
of just setting a rule in the convention plugin and expecting consuming
projects to be able to override it using Kover's DSL would not work as
desired.

Even though there's only a single property, structure the extension
hierarchically by function area in an attempt to allow for future
expansion.