Add configuration property for excluding classes from code coverage
This is a more-or-less transparent wrapper around Kover's equivalent
property, and using its DSL is not particularly onerous, so the primary
benefit of this is simply API consistency: a consuming project should
not have to configure code coverage settings in two places.  This does
also shield consuming projects from knowing about Kover at all, though,
so there may also be some benefit in case we eventually want to change
code coverage plugins.