Create CI/CD workflow for publishing to GitHub Packages
This is my first attempt to use GitHub Packages, so I am not expecting
that this will work on the first try.  Nevertheless, I would rather
debug it post-merge, in CI/CD, rather than trying to also figure out how
to get it to work locally so I can debug it ahead of time.

Because we want the version specified for a CI/CD release to override
the "local-SNAPSHOT" version used for local development, we must either
move the local definition of the `version` property to a different
location (since specifying it directly in `build.gradle.kts` is
higher-priority than a command-line specification), or else have the
CI/CD version be specified indirectly (say, as an environment variable)
and explicitly construct the version in `build.gradle.kts` using the
Elvis operator.  I've opted for the former, as it seems more
straightforward, and moved `version` to a `gradle.properties` file.
Move the `group` too, since it's related and not complicated in its
handling.

Note that the version identifier for the published artifact will be
taken directly from the GitHub release creating it.  This implies that
release tags should be exactly identical to the semantic versions they
name, and in particular should not use the "v" prefix that is sometimes
suggested.  A quick look at some authoritative Kotlin projects didn't
turn up any that use the "v", so this seems to be the majority view.